set dateModified on save
Also move DatabaseCollection into its own file, for easier maintainability / modularity.
